app/requirements.txt
langchain
fastapi
uvicorn
langchain
langchain_community
langchain_openai
langchain_experimental
python-dotenv